transcript json will fill all missing time
Now the transcript json fills all missing timestamps in maintranscribe using utils. fill_all_missing_times function removed from json_to_srt since the transcript json already has the times filled.
