+++ src/transcribe/maintranscribe.py
@@ -1,9 +1,10 @@
-import os
+import json, os
 from .. import config
 import whisperx
 from .converters.json_to_srt import convert_to_srt
 from .converters.json_to_transcript import convert_to_transcript
 from .converters.json_to_words import filter_words
+from .converters import utils
 
 def get_filename(filepath):
     filename = os.path.basename(filepath)
@@ -26,6 +27,8 @@ def transcribe(audio_file, speakers: int, language="en", device="cuda", model_na
         diarize_segments = diarize_model(audio_file, min_speakers=1, max_speakers=speakers)
         result = whisperx.assign_word_speakers(diarize_segments, result)
         print("Diarized")
+    
+    result = utils.fill_all_missing_times(result["segments"])
 
     abs_transcript_folder = os.path.abspath(config.transcript_folder)
     config.json_path = os.path.join(abs_transcript_folder, f"{get_filename(audio_file)}.json")
@@ -38,10 +41,8 @@ def transcribe(audio_file, speakers: int, language="en", device="cuda", model_na
     print("srt path: ", config.srt_path)
     print("trs path: ", config.trs_path) """
 
-    import json
-
     with open(config.json_path, "w") as outfile:
-        json.dump(result["segments"], outfile)
+        json.dump(result, outfile, indent=4)
 
     convert_to_srt(config.json_path, config.srt_path)
     convert_to_transcript(config.json_path, config.trs_path, speaker=speakers)
